Cwdoe 564 api side sorting (#104)
If changes to a ScenarioEvent change the DeltaSeconds or the GroupOrder properties, the sorting of the events within the MSEL may need to be changed. This MR implements this functionality and changes the return type for both creating and updating events to be a list since multiple events can be affected when ordering changes are triggered.

This MR also includes a DB migration that only corrects any ordering errors on an old version of the schema where events within a group (i.e. events that happen at the same time) are not ordered by GroupOrder already.


---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>